+++ sdmap/src/sdmap.ext.Dapper/SdmapExtensions.cs
@@ -1,6 +1,4 @@
-﻿using Dapper;
-using sdmap.Compiler;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Data;
 using System.IO;
@@ -9,7 +7,7 @@ using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace sdmap.ext
+namespace sdmap.ext.Dapper
 {
     public static partial class SdmapExtensions
     {